demo: move location follow onto the overlay stack
Replace the My location demo with a follow button on the shared control
column and the engine selector in settings.
