+++ demo-app/common/src/commonMain/composeResources/drawable/my_location_24px.xml
@@ -1,0 +1,13 @@
+<vector
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  android:width="24dp"
+  android:height="24dp"
+  android:viewportWidth="960"
+  android:viewportHeight="960"
+  android:tint="?attr/colorControlNormal"
+>
+  <path
+    android:fillColor="#ffffff"
+    android:pathData="M440,918L440,838Q315,824 225.5,734.5Q136,645 122,520L42,520L42,440L122,440Q136,315 225.5,225.5Q315,136 440,122L440,42L520,42L520,122Q645,136 734.5,225.5Q824,315 838,440L918,440L918,520L838,520Q824,645 734.5,734.5Q645,824 520,838L520,918L440,918ZM480,760Q596,760 678,678Q760,596 760,480Q760,364 678,282Q596,200 480,200Q364,200 282,282Q200,364 200,480Q200,596 282,678Q364,760 480,760ZM480,640Q414,640 367,593Q320,546 320,480Q320,414 367,367Q414,320 480,320Q546,320 593,367Q640,414 640,480Q640,546 593,593Q546,640 480,640ZM480,560Q513,560 536.5,536.5Q560,513 560,480Q560,447 536.5,423.5Q513,400 480,400Q447,400 423.5,423.5Q400,447 400,480Q400,513 423.5,536.5Q447,560 480,560ZM480,480Q480,480 480,480Q480,480 480,480Q480,480 480,480Q480,480 480,480Q480,480 480,480Q480,480 480,480Q480,480 480,480Z"
+  />
+</vector>
